簡易的な聴いた動画削除機能を実装。投稿者本人のみ削除できるようにする必要がある

+++ app/express/src/application/comment/CommentApplicationService.ts
@@ -71,6 +71,10 @@ export class CommentApplicationService {
 
       const comment = await this.commentRepository.findOneById(commentId);
 
+      if (!comment) {
+        throw new Error('deleteしようとしたcommentはすでに存在ません');
+      }
+
       await this.commentRepository.delete(comment);
 
       return { ok: true, error: null };
